+++ packages/jest-cli/src/reporters/notify_reporter.js
@@ -31,7 +31,7 @@ export default class NotifyReporter extends BaseReporter {
     startRun: (globalConfig: GlobalConfig) => *,
     context: TestSchedulerContext,
   ) {
-    super();
+    super(globalConfig);
     this._globalConfig = globalConfig;
     this._startRun = startRun;
     this._context = context;